requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
PyJWT==1.7.1
pytest==7.4.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
